refactor: 자동 로그인 여부 로직 수정
