+++ feature/src/main/java/com/sopt/now/feature/MainActivity.kt
@@ -4,9 +4,11 @@ import android.view.MotionEvent
 import android.view.View
 import android.view.inputmethod.InputMethodManager
 import androidx.activity.OnBackPressedCallback
+import androidx.activity.viewModels
 import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.NavController
+import androidx.navigation.Navigation
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.ui.setupWithNavController
@@ -25,9 +27,12 @@ class MainActivity : BindingActivity<ActivityMainBinding>(R.layout.activity_main
     private var backPressedTime = 0L
     private val backPressedFlow = MutableSharedFlow<Unit>()
 
+    private val viewModel: MainViewModel by viewModels()
+
     override fun initView() {
         initMainBottomNavigation()
         initBackDoublePressed()
+        observeAutoLogin()
     }
 
     private fun initMainBottomNavigation() {
@@ -84,6 +89,16 @@ class MainActivity : BindingActivity<ActivityMainBinding>(R.layout.activity_main
             backPressedTime = currentTime
             snackBar(binding.root, getString(R.string.main_back_once_pressed_exit))
         }
+    }
+
+    private fun observeAutoLogin() {
+        viewModel.autoLoginState.flowWithLifecycle(lifecycle).onEach { isAutoLogin ->
+            val navController = Navigation.findNavController(this, R.id.fcv_home)
+            when (isAutoLogin) {
+                true -> navController.navigate(R.id.fragment_home)
+                false -> navController.navigate(R.id.fragment_login)
+            }
+        }.launchIn(lifecycleScope)
     }
 
     override fun dispatchTouchEvent(ev: MotionEvent): Boolean {
